// src/types/ITheme.ts
export interface ITheme {
  colors: {
    fontColor: string;
    mainBackground: string;
    mainBorder: string;
    boxShadow: string;
    selectBackground: string;
    selectFontColor: string;
    headerFooterBackground: string;
    modalBackground: string;
    toggleBackground: string;
    filmsCardBackground: string;
    toggleBackgroundChecked: string;
    toggleCircle: string;
    toggleBorder: string;
    hoverBoxShadow: string;
  };
}